+++ test/codemelted_test.dart
@@ -45,10 +45,168 @@ void main() {
   // This is called once after EVERY test
   tearDown(() {});
 
-  group("codemelted.math Tests", () {
-    test("Temperature Conversion Validation", () {
-      var answer = codemelted.math.calculate(
-        formula: Formula_t.temperature_celsius_to_fahrenheit,
+  // --------------------------------------------------------------------------
+  // [JSON Use Case] ----------------------------------------------------------
+  // --------------------------------------------------------------------------
+
+  group("codemelted.json Tests", () {
+    test("CArray Validation", () {
+      var listenerCount = 0;
+      onDataChanged() => listenerCount += 1;
+
+      var array = codemelted.json.createArray();
+      array.addListener(listener: onDataChanged);
+      expect(array.isEmpty, isTrue);
+
+      // Fire listener
+      array.add(2);
+      array.notifyAll();
+      expect(array.isNotEmpty, isTrue);
+      expect(listenerCount, equals(1));
+
+      // Don't fire listener
+      array.insert(0, "hello");
+      expect(array.length, equals(2));
+      expect(listenerCount, equals(1));
+
+      // Now validate index works as expected.
+      array.insert(1, true);
+      array.notifyAll();
+      expect(array.length, equals(3));
+      expect(listenerCount, equals(2));
+      expect(array[1], isTrue);
+      expect(array[2], equals(2));
+      expect(array[0], equals("hello"));
+
+      // Make a copy and prove it is a copy
+      var arrayCopy = array.copy();
+      arrayCopy.add("another string");
+      expect(array.length != arrayCopy.length, isTrue);
+
+      var stringified = arrayCopy.stringify();
+      expect(stringified!.contains("another"), isTrue);
+
+      // Now validate the parse clears the data of the original array
+      // and load the copies data as if it were serialized data read from
+      // storage to be used.
+      var success = array.parse(data: stringified);
+      expect(success, isTrue);
+    });
+
+    test("CObject Validation", () {
+      var listenerCount = 0;
+      onDataChanged() => listenerCount += 1;
+
+      var obj = codemelted.json.createObject();
+      obj.addListener(listener: onDataChanged);
+      expect(obj.isEmpty, isTrue);
+
+      obj.set<String>(key: "test", value: "test", notify: true);
+      expect(listenerCount, equals(1));
+      expect(obj.isNotEmpty, isTrue);
+      expect(obj.get<String>(key: "test"), equals("test"));
+
+      obj.set<bool?>(key: "is_good", value: null);
+      expect(obj.get<bool?>(key: "is_good"), isNull);
+      expect(listenerCount, equals(1));
+
+      obj.set<CArray>(key: "anArray", value: [1, true, "three"], notify: true);
+      expect(listenerCount, equals(2));
+      expect(obj.get<CArray>(key: "anArray").length, equals(3));
+
+      var copy = obj.copy();
+      copy.set<int>(key: "newKey", value: 2);
+      expect(listenerCount, equals(2));
+      expect(copy.entries.length != obj.entries.length, isTrue);
+
+      var stringified = obj.stringify();
+      expect(stringified, isNotNull);
+
+      var success = obj.parse(initData: stringified!);
+      expect(success, isTrue);
+    });
+
+    test("asXXX() Validation", () {
+      // Array validation
+      expect(codemelted.json.asArray(data: "{}"), isNull);
+      var array = codemelted.json.createArray(
+        initData: [1, true, "awesome", null],
+      );
+      var stringified = array.stringify();
+      expect(codemelted.json.asArray(data: stringified!), isNotNull);
+
+      // Bool validation
+      expect(codemelted.json.asBool(data: "42"), isFalse);
+      expect(codemelted.json.asBool(data: "1"), isTrue);
+      expect(codemelted.json.asBool(data: "yes"), isTrue);
+      expect(codemelted.json.asBool(data: "no"), isFalse);
+
+      // double validation
+      expect(codemelted.json.asDouble(data: "a"), isNull);
+      expect(codemelted.json.asDouble(data: "42.2"), isNotNull);
+      expect(codemelted.json.asDouble(data: "42"), isNotNull);
+
+      // int validation
+      expect(codemelted.json.asInt(data: "a"), isNull);
+      expect(codemelted.json.asInt(data: "42.2"), isNull);
+      expect(codemelted.json.asInt(data: "42"), isNotNull);
+
+      // obj validation
+      expect(codemelted.json.asObject(data: "[]"), isNull);
+      var obj = codemelted.json.createObject(initData: {
+        "one": 1,
+        "two": "two",
+        "three": null,
+      });
+      stringified = obj.stringify();
+      expect(codemelted.json.asObject(data: stringified!), isNotNull);
+    });
+
+    test("checkXXX() / tryXXX() Validation", () {
+      var obj = codemelted.json.createObject(initData: {
+        "one": 1,
+        "two": "two",
+        "three": null,
+      });
+
+      expect(codemelted.json.checkHasProperty(obj: obj, key: "one"), isTrue);
+      expect(codemelted.json.checkHasProperty(obj: obj, key: "4"), isFalse);
+      try {
+        codemelted.json.tryHasProperty(obj: obj, key: "4");
+        fail("should throw exception");
+      } catch (ex) {
+        expect(ex, isA<String>());
+      }
+
+      expect(codemelted.json.checkType<String>(data: 42), isFalse);
+      expect(codemelted.json.checkType<String>(data: "42"), isTrue);
+      try {
+        codemelted.json.tryType<String>(data: 42);
+        fail("should throw exception");
+      } catch (ex) {
+        expect(ex, isA<String>());
+      }
+
+      expect(codemelted.json.checkValidUrl(data: "https://codemelted.com"),
+          isTrue);
+      expect(codemelted.json.checkValidUrl(data: "j;,ht:/b;l42"), isFalse);
+      try {
+        codemelted.json.tryValidUrl(data: "j;,ht:/b;l42");
+        fail("should throw exception");
+      } catch (ex) {
+        expect(ex, isA<String>());
+      }
+    });
+  });
+
+  // --------------------------------------------------------------------------
+  // [NPU Use Cases] ----------------------------------------------------------
+  // --------------------------------------------------------------------------
+
+  group("codemelted.npu Tests", () {
+    test("math - Temperature Conversion Validation", () {
+      var answer = codemelted.npu.math(
+        formula: CFormula_t.temperature_celsius_to_fahrenheit,
         arg1: 0.0,
       );
       expect(answer, moreOrLessEquals(32.0));