+++ CHANGELOG.md
@@ -1,3 +1,5 @@
-## 0.0.0
+## 0.1.1 (Last Modified 2024-12-27)
 
-* TBD
+* Cleanup of a lot of documentation
+* Have the NPU WASM module hooked up with one formula.
+* Completed the documentation and testing of the codemelted.json namespace.